# bit positions over a whole stream wrapped an int and crashed the decoders
A decoder that walks the compressed stream one bit at a time held its position in
an int. That stops fitting at 2^28 bytes, and the guard reads a wrapped negative
position as in range, so the next array access threw IndexOutOfRangeException.

Unary coding reached it first: it emits up to 256 bits per input byte, so 16.7 MB
of input already codes to more than 2^31 bits. Measured, the round trip threw at
16,777,216 bytes and now succeeds through 67,108,864.

Widened to long in Unary, Elias Gamma, Elias Delta, Levenshtein, Golomb, Tunstall,
Shannon-Fano, Exp-Golomb, Golomb-Rice, FSE, LZS, LZWL, LZHAM, SQX, IBM 842, ZIP
Implode and the Brotli bit reader. Below the wrap point the widened arithmetic is
exactly the old arithmetic, so output is unchanged.

+++ Compression.Core/Dictionary/Brotli/BrotliBitReader.cs
@@ -35,7 +35,7 @@ internal sealed class BrotliBitReader {
   public int BytePosition => this._bytePos - (this._bitsAvailable >> 3);
 
   /// <summary>Gets the exact bit position in the stream.</summary>
-  public int BitPosition => this._bytePos * 8 - this._bitsAvailable;
+  public long BitPosition => (long)this._bytePos * 8 - this._bitsAvailable;
 
   /// <summary>Debug: dump buffer state.</summary>
   public string DebugState => $"bytePos={this._bytePos} bitsAvail={this._bitsAvailable} buf=0x{this._bitBuffer:X16} bitPos={this.BitPosition}";
