+++ Compression.Core/Entropy/EliasGammaBuildingBlock.cs
@@ -46,7 +46,9 @@ public sealed class EliasGammaBuildingBlock : IBuildingBlock {
 
     var src = data[4..].ToArray();
     var result = new byte[originalSize];
-    var bitIndex = 0;
+    // Elias Gamma emits up to 17 bits per input byte, so the bit position passes
+    // int.MaxValue while the stream is still far short of Array.MaxLength.
+    var bitIndex = 0L;
 
     for (var i = 0; i < originalSize; i++)
       result[i] = (byte)(DecodeGamma(src, ref bitIndex) - 1);
@@ -72,7 +74,7 @@ public sealed class EliasGammaBuildingBlock : IBuildingBlock {
       writer.WriteBit((value >> i) & 1);
   }
 
-  private static int DecodeGamma(byte[] data, ref int bitIndex) {
+  private static int DecodeGamma(byte[] data, ref long bitIndex) {
     // Count leading zero-bits.
     var n = 0;
     while (ReadBit(data, ref bitIndex) == 0)
@@ -86,10 +88,10 @@ public sealed class EliasGammaBuildingBlock : IBuildingBlock {
     return value;
   }
 
-  private static int ReadBit(byte[] data, ref int bitIndex) {
+  private static int ReadBit(byte[] data, ref long bitIndex) {
     if (bitIndex / 8 >= data.Length)
       throw new InvalidDataException("Unexpected end of Elias Gamma bitstream.");
-    var bit = (data[bitIndex / 8] >> (7 - (bitIndex % 8))) & 1;
+    var bit = (data[bitIndex / 8] >> (7 - (int)(bitIndex % 8))) & 1;
     bitIndex++;
     return bit;
   }
